Added PyEE Files
Version 0.1

This program is intended to streamline impedance calculations for electrical engineers.

Impedances come in standardized values, but an electrical engineer might need an intermediate value of impedance. That can be achieved through placing impedances in parallel, but the calculations to determine what combination of which impedances can become cumbersome and tedious. This program utilizes looping and basic principles of electrical engineering to vastly expedite that process by calculating every possible equivalent parallel duo of impedance in a given impedance list.

-Example-
given a list of impedances "1,2,3,4,5" returns 1||1, 1||2, up to 1||5; 2||2, 2||3, up to 2||5; so on and so forth.

- Utilizes the tkinter GUI package for convenient usage.
- Can calculate the value of an n-combination of parallel impedances. This is useful if one wants to quickly calculate the equivalent impedance of an n'th number of parallel impedances.

-Example- 
Given "1,2,3,4,5" returns the value of "1||2||3||4||5"

- Works for capacitors in series, resistors in parallel, and inductors in parallel.
- Does NOT (yet) work for complex equivalent impedance calculation, but can be used for the real and reactive components individually.